demonstrate the problem
Click on the bubbles. Some clicks work properly (highlight the
bubble that was clicked), some clicks highlight the wrong bubble,
and some clicks do nothing. Note errors in the de-highlight process
as well - clicking on a separate bubble sometimes de-highlights
the first.

I also refactored MainActivity. Previously it was painful to add
a new activity, especially if it is not at the bottom of the list.
The new implementation is cleaner - WISYWIG and easier to maintain.